requirements.txt
ultralytics>=8.2.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
pandas>=2.0.0
tqdm>=4.66.0
requests>=2.31.0
scipy>=1.10.0